refactor(adapter-interfaces): extract shared base64url helper for cursor codec
