+++ packages/adapter-interfaces/src/utils/index.ts
@@ -2,4 +2,5 @@ export * from "./user-id";
 export * from "./passthrough";
 export * from "./connection-attributes";
 export * from "./guards";
+export * from "./base64url";
 export * from "./cursor";
